check if pi is enqueued to run in the future before running it in celery w/ burnettk
