+++ spiffworkflow-backend/src/spiffworkflow_backend/routes/process_instances_controller.py
@@ -662,7 +662,7 @@ def _process_instance_run(
         ProcessInstanceTmpService.add_event_to_process_instance(process_instance, "process_instance_force_run")
         if not queue_process_instance_if_appropriate(
             process_instance, execution_mode=execution_mode
-        ) and not ProcessInstanceQueueService.is_enqueued_to_run_in_the_future(process_instance):
+        ) and not ProcessInstanceTmpService.is_enqueued_to_run_in_the_future(process_instance):
             execution_strategy_name = None
             if execution_mode == ProcessInstanceExecutionMode.synchronous.value:
                 execution_strategy_name = "greedy"
